+++ lib/screens/login.screen.dart
@@ -133,122 +133,18 @@ class _LoginScreenState extends State<LoginScreen> {
             ),
           ),
 
-          /// workload box
-          _workloadText(context, this.loginPayload!.workloadText, size),
-
-          SizedBox(height: size.height * 0.05),
+          /// profundity text
+          Padding(
+            padding: const EdgeInsets.only(bottom: 16),
+            child: WorkloadMarkdown(
+              workload: this.loginPayload!.workloadText,
+            ),
+          ),
 
           /// formulario (falta aplicar backend)
           _loginForm(context),
         ],
       ],
-    );
-  }
-
-  _workloadText(BuildContext context, String workloadText, Size size) {
-    return Container(
-      width: double.infinity,
-      margin: EdgeInsets.symmetric(horizontal: 40),
-      padding: EdgeInsets.symmetric(vertical: size.width * 0.1),
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.all(Radius.circular(7.0) //
-            ),
-        color: Colors.white,
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black,
-            offset: Offset(0.0, 0.7), //(x,y)
-            blurRadius: 1.0,
-          ),
-        ],
-      ),
-      child: Column(
-        children: [
-          Container(
-            padding: EdgeInsets.symmetric(horizontal: size.width * 0.1),
-            child: Text(
-              "Carga horaria estimada para cada docente por ciudad:",
-              style: korolevFont.headline5?.copyWith(
-                  fontSize: 20,
-                  color: Colors.black,
-                  fontWeight: FontWeight.bold),
-              textAlign: TextAlign.center,
-            ),
-          ),
-          SizedBox(
-            height: 10,
-          ),
-          Text(
-            "10 horas",
-            style: korolevFont.headline6?.apply(
-              fontSizeFactor: 0.9,
-              color: ColorsApp.backgroundRed,
-            ),
-          ),
-          SizedBox(
-            height: 10,
-          ),
-          Container(
-            padding: EdgeInsets.symmetric(horizontal: size.width * 0.05),
-            child: Text(
-              "Se trata de 4 horas para el consumo asíncrono de contenidos y 6 horas para el tiempo de producción y gamifiación",
-              style: korolevFont.headline5?.apply(
-                fontSizeFactor: 0.6,
-                color: Colors.black,
-              ),
-              textAlign: TextAlign.center,
-            ),
-          ),
-          SizedBox(
-            height: 30,
-          ),
-          Container(
-            padding: EdgeInsets.symmetric(horizontal: size.width * 0.1),
-            child: Text(
-              "Tiempo total del viaje 2222",
-              style: korolevFont.headline5?.copyWith(
-                  fontSize: 20,
-                  color: Colors.black,
-                  fontWeight: FontWeight.bold),
-              textAlign: TextAlign.center,
-            ),
-          ),
-          SizedBox(
-            height: 10,
-          ),
-          Text(
-            "120 horas",
-            style: korolevFont.headline6?.apply(
-              fontSizeFactor: 0.9,
-              color: ColorsApp.backgroundRed,
-            ),
-          ),
-          SizedBox(
-            height: 15,
-          ),
-          Container(
-            padding: EdgeInsets.symmetric(horizontal: size.width * 0.05),
-            child: Text(
-              "para una duración máxima de 45 días",
-              style: korolevFont.headline5?.apply(
-                fontSizeFactor: 0.6,
-                color: Colors.black,
-              ),
-              textAlign: TextAlign.center,
-            ),
-          ),
-          SizedBox(
-            height: 20,
-          ),
-          Text(
-            "A disfrutar...",
-            style: korolevFont.headline6?.apply(
-              fontSizeFactor: 0.9,
-              color: ColorsApp.backgroundRed,
-            ),
-          ),
-        ],
-      ),
     );
   }
 
@@ -308,7 +204,6 @@ class _LoginScreenState extends State<LoginScreen> {
   ///campo de usuario
   _userField(BuildContext context) {
     return Container(
-      margin: EdgeInsets.symmetric(horizontal: 40),
       height: 40,
       decoration: BoxDecoration(
         color: Colors.white,
@@ -337,7 +232,6 @@ class _LoginScreenState extends State<LoginScreen> {
   ///campo de contraseña
   _passwordField(BuildContext context) {
     return Container(
-      margin: EdgeInsets.symmetric(horizontal: 40),
       height: 40,
       decoration: BoxDecoration(
         color: Colors.white,
@@ -367,7 +261,6 @@ class _LoginScreenState extends State<LoginScreen> {
     double buttonWidth = MediaQuery.of(context).size.width;
     return Container(
       width: buttonWidth,
-      margin: EdgeInsets.symmetric(horizontal: 40),
       child: ElevatedButton(
         style: ElevatedButton.styleFrom(
           primary: ColorsApp.backgroundBottomBar,
@@ -386,3 +279,5 @@ class _LoginScreenState extends State<LoginScreen> {
     );
   }
 }
+
+
