+++ lib/shared/widgets/markdown.widget.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_markdown/flutter_markdown.dart' as md;
+import 'package:flutter_markdown/flutter_markdown.dart' as fmd;
+import 'package:lab_movil_2222/themes/colors.dart';
+import 'package:markdown/markdown.dart' as md;
 
 @Deprecated('use the Markdown2222 instead')
 typedef Markdown = Markdown2222;
@@ -14,9 +16,179 @@ class Markdown2222 extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return md.MarkdownBody(
+    return fmd.MarkdownBody(
       data: this.data,
-      styleSheet: md.MarkdownStyleSheet.fromTheme(Theme.of(context)),
+      styleSheet: fmd.MarkdownStyleSheet.fromTheme(Theme.of(context)),
+    );
+  }
+}
+
+/// Markdown content adapted to implement design guides, The team content is
+/// mainly centered
+class TeamContentMarkdown extends StatelessWidget {
+  const TeamContentMarkdown({
+    Key? key,
+    required this.teamContent,
+  }) : super(key: key);
+
+  final String teamContent;
+
+  @override
+  Widget build(BuildContext context) {
+    final TextTheme textTheme = Theme.of(context).textTheme;
+
+    return fmd.MarkdownBody(
+      data: this.teamContent,
+      builders: {
+        'h2': _MarkdownCenterText(),
+        'h3': _MarkdownCenterText(padding: EdgeInsets.only(top: 35)),
+        'h4': _MarkdownCenterText(padding: EdgeInsets.only(top: 10)),
+        'p': _MarkdownCustomCreators(),
+      },
+      styleSheet: fmd.MarkdownStyleSheet(
+        h2: textTheme.headline4,
+        h3: textTheme.headline6,
+        h4: textTheme.caption,
+        p: textTheme.subtitle1,
+      ),
+    );
+  }
+}
+
+///
+///
+///
+class WorkloadMarkdown extends StatelessWidget {
+  final String workload;
+
+  const WorkloadMarkdown({
+    Key? key,
+    required this.workload,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final TextTheme textTheme = Theme.of(context).primaryTextTheme;
+    final Size size = MediaQuery.of(context).size;
+
+    return Container(
+      decoration: BoxDecoration(
+        color: ColorsApp.white,
+        borderRadius: BorderRadius.circular(8),
+      ),
+      padding: EdgeInsets.all(size.width * 0.1),
+      child: fmd.MarkdownBody(
+        data: this.workload,
+        builders: {
+          'h2': _MarkdownCenterText(padding: EdgeInsets.symmetric(vertical: 8)),
+          'h3': _MarkdownCenterText(padding: EdgeInsets.symmetric(vertical: 4)),
+          'p': _MarkdownCenterText(padding: EdgeInsets.symmetric(vertical: 4)),
+        },
+        styleSheet: fmd.MarkdownStyleSheet(
+          h2: textTheme.subtitle1,
+          h3: textTheme.subtitle1!.apply(color: ColorsApp.red),
+          p: textTheme.bodyText2,
+        ),
+      ),
+    );
+  }
+}
+
+/// utility to center markdown text
+class _MarkdownCenterText extends fmd.MarkdownElementBuilder {
+  final EdgeInsets? padding;
+
+  _MarkdownCenterText({
+    this.padding,
+  });
+
+  @override
+  Widget? visitText(md.Text element, TextStyle? preferredStyle) {
+    return Container(
+      padding: this.padding,
+      width: double.infinity,
+      child: Text(
+        element.text,
+        style: preferredStyle,
+        textAlign: TextAlign.center,
+      ),
+    );
+  }
+}
+
+/// utility for custom layout of teams page
+class _MarkdownCustomCreators extends fmd.MarkdownElementBuilder {
+  @override
+  Widget? visitText(md.Text element, TextStyle? preferredStyle) {
+    final List<String> names =
+        element.text.split(',').map((e) => e.trim()).toList();
+
+    final List<Widget> colItems = [];
+
+    for (int i = 0; i < names.length; i += 2) {
+      final String nameOne = names[i];
+      final String? nameTwo = (i + 1 >= names.length) ? null : names[i + 1];
+
+      final Row row = Row(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+          /// add first name
+          Expanded(
+            child: Container(
+              width: double.infinity,
+              child: Text(
+                nameOne,
+                style: preferredStyle,
+                maxLines: 1,
+                overflow: TextOverflow.ellipsis,
+                textAlign: nameTwo == null ? TextAlign.center : TextAlign.right,
+              ),
+            ),
+          ),
+
+          /// add second name then found
+          if (nameTwo != null) ...[
+            SizedBox(
+              width: 10,
+            ),
+            Container(
+              width: 8,
+              height: 8,
+              decoration: BoxDecoration(
+                borderRadius: BorderRadius.circular(10),
+                color: Colors.white,
+              ),
+            ),
+            SizedBox(
+              width: 10,
+            ),
+            Expanded(
+              child: Text(
+                nameTwo,
+                style: preferredStyle,
+                maxLines: 1,
+                overflow: TextOverflow.ellipsis,
+                textAlign: TextAlign.left,
+              ),
+            ),
+          ]
+        ],
+      );
+
+      colItems.add(row);
+
+      /// add spacer between lines
+      if (i + 1 < names.length)
+        colItems.add(
+          Container(
+            height: 5,
+          ),
+        );
+    }
+
+    return Column(
+      children: colItems,
+      crossAxisAlignment: CrossAxisAlignment.stretch,
     );
   }
 }
